I think I got most/all of the strict skill leveling functionality done.
The only problem right now is that decreasing levels is extremely slow. The simplest remedy would be to start storing the data-lvl to store. Could also do things like keep a store for the Skill.NeedSP and Skill.ParentSkills... but technically there aren't many skills  to make a huge performance boost.