Upgraded to alpha 6 of bootstrap. Now need a row.
I MIGHT split up the job parts, might not. Not sure yet.

I have an idea for a layout/design. Not too sure if it'll work. It'll look something like:

[ 1 ] [ 2 ]
[ 3 ]

or
[ 1 ] [ 2 ]
[ 3 ] [ 4 ]

Where 1 = base, 2 = first ascension, 3 = second ascension, and 4 = third (awakened). And to the right side of all of this will be the description.

Ideally I would like all of the trees shown on the page, but maybe it's better to just show 1 at a time like in YADNSS. I'm still thinking it over. Hopefully will have a decision within 24 hours.